# specs/001-graphrag-chatbot-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: GraphRAG Chatbot Platform API
  description: PDF 기반 GraphRAG 챗봇 플랫폼 API
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
...
paths:
...
  /chat/{accessUrl}:
    get:
      tags: [Chat]
      summary: 챗봇 정보 조회 (공개)
      operationId: getChatbotPublic
      parameters:
        - name: accessUrl
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 챗봇 공개 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatbotPublic'
        '404':
          $ref: '#/components/responses/NotFound'
  /chat/{accessUrl}/sessions:
    post:
      tags: [Chat]
      summary: 대화 세션 생성
      operationId: createSession
      parameters:
        - name: accessUrl
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: 세션 생성 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          $ref: '#/components/responses/NotFound'
  /chat/{accessUrl}/sessions/{sessionId}/messages:
    get:
      tags: [Chat]
      summary: 대화 히스토리 조회
      operationId: getMessages
      parameters:
        - name: accessUrl
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: 메시지 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'

    post:
      tags: [Chat]
      summary: 질문 전송 (스트리밍 응답)
      operationId: sendMessage
      parameters:
        - name: accessUrl
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: SSE 스트리밍 응답
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  SSE 이벤트 스트림:
                  - event: token, data: "토큰 텍스트"
                  - event: sources, data: JSON 출처 배열
                  - event: done, data: ""
                  - event: error, data: "오류 메시지"
  /chat/{accessUrl}/sessions/{sessionId}/stop:
    post:
      tags: [Chat]
      summary: 응답 생성 중단
      operationId: stopGeneration
      parameters:
        - name: accessUrl
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: 중단 완료

  # ============ Stats ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    ChatbotPublic:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        persona:
          type: object
          properties:
            greeting:
              type: string
...
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    MessageRole:
      type: string
      enum: [user, assistant]
    Source:
      type: object
      properties:
        document_id:
          type: string
          format: uuid
        document_name:
          type: string
        page:
          type: integer
        section:
          type: string
        relevance_score:
          type: number
          format: float
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/MessageRole'
        content:
          type: string
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
        created_at:
          type: string
          format: date-time
    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 10000

    # ---- Stats ----
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object